fix(#690): move sync_bridge from core/ to lib/
sync_bridge.py is a pure stdlib utility (asyncio, threading, logging)
with zero nexus dependencies — belongs in the tier-neutral lib/ package,
not in core/. Updated all 50 import sites across backends, services,
cli, fuse, mcp, and tests.
