+++ src/nexus/backends/gmail_connector.py
@@ -390,7 +390,7 @@ class GmailConnectorBackend(
             )
 
         # Get valid access token from TokenManager (auto-refreshes if expired)
-        from nexus.core.sync_bridge import run_sync
+        from nexus.lib.sync_bridge import run_sync
 
         try:
             # Default to 'default' zone if not specified to match mount configurations
