+++ src/nexus/core/nexus_fs.py
@@ -5722,7 +5722,7 @@ class NexusFS(  # type: ignore[misc]
         Returns:
             Result of the coroutine
         """
-        from nexus.core.sync_bridge import run_sync
+        from nexus.lib.sync_bridge import run_sync
 
         return run_sync(coro)
