+++ src/nexus/fuse/ops/_shared.py
@@ -569,7 +569,7 @@ def resolve_io_profile(ctx: FUSESharedContext, path: str) -> str:
         ctx._io_profile_loaded = True  # type: ignore[attr-defined]
         try:
             mount_svc = cast(Any, ctx.nexus_fs).mount_service
-            from nexus.core.sync_bridge import run_sync as _run_sync
+            from nexus.lib.sync_bridge import run_sync as _run_sync
 
             mounts = _run_sync(mount_svc.list_mounts())
             ctx._io_profile_mounts = sorted(  # type: ignore[attr-defined]
