+++ src/nexus/services/ace/consolidation.py
@@ -50,7 +50,7 @@ def _run_coroutine(coro: Any) -> Any:
     Returns:
         The coroutine's return value.
     """
-    from nexus.core.sync_bridge import run_sync
+    from nexus.lib.sync_bridge import run_sync
 
     return run_sync(coro)
