+++ src/nexus/services/ace/learning_loop.py
@@ -215,7 +215,7 @@ class LearningLoop:
         Returns:
             Execution results
         """
-        from nexus.core.sync_bridge import run_sync
+        from nexus.lib.sync_bridge import run_sync
 
         return run_sync(
             self.execute_with_learning_async(
@@ -287,7 +287,7 @@ class LearningLoop:
         Returns:
             List of re-learning results
         """
-        from nexus.core.sync_bridge import run_sync
+        from nexus.lib.sync_bridge import run_sync
 
         return run_sync(self.process_relearning_queue_async(limit))
