+++ src/nexus/services/memory/coref_resolver.py
@@ -170,7 +170,7 @@ Resolved text:"""
         Returns:
             CorefResult with resolved text.
         """
-        from nexus.core.sync_bridge import run_sync
+        from nexus.lib.sync_bridge import run_sync
 
         result: CorefResult = run_sync(self.resolve_async(text, context, entity_hints))
         return result
