+++ src/nexus/services/memory/temporal_resolver.py
@@ -159,7 +159,7 @@ Resolved:"""
         Returns:
             TemporalResult with resolved text.
         """
-        from nexus.core.sync_bridge import run_sync
+        from nexus.lib.sync_bridge import run_sync
 
         return run_sync(self.resolve_async(text, reference_time, context))
