+++ tests/e2e/server/test_sync_bridge_fastapi_e2e.py
@@ -22,7 +22,7 @@ import httpx
 import pytest
 
 from nexus.backends.local import LocalBackend
-from nexus.core.sync_bridge import shutdown_sync_bridge
+from nexus.lib.sync_bridge import shutdown_sync_bridge
 from nexus.storage.record_store import SQLAlchemyRecordStore
